+++ NEWS.md
@@ -1,12 +1,14 @@
-# gatpkg v2.0.0
+# gatpkg v2.0.1
 
 * Added check for operating system in `locateGATshapefile()`.
+* Removed dependency on spdep and fixed associated bugs.
+* Removed function `aggregateGATnb()`.
 
 # gatpkg v2.0.0
 
-* Revised `writeGATkml` so that it also writes to KMZ.
+* Revised `writeGATkml()` so that it also writes to KMZ.
 * Removed dependencies on rgeos, lwgeom, rgdal, foreign, tcltk2, plotkml, and sp.
-* Removed `checkGATshapefilesize` and fixed bugs. 
+* Removed `checkGATshapefilesize()` and fixed bugs. 
 * Added advanced options to the settings file and log.
 * Fixed several bugs that broke GAT.
 * Revised and standardized all pop-up fonts and colors except the pre-defined progress bar and file-selection window.